reorganized test directory
